Lansitec decoders, add the unit test for the payload.
