+++ decoders/connector/lansitec/nb-iot-macro-bluetooth-gateway/connector.jsonc
@@ -6,7 +6,7 @@
   },
   "versions": {
     "v1.0.0": {
-      "src": "./v1.0.0/payload.js",
+      "src": "./v1.0.0/payload.ts",
       "manifest": "./v1.0.0/payload-config.jsonc"
     }
   }
