+++ decoders/connector/lansitec/socket-sync-bluetooth-gateway/v1.0.0/payload.ts
@@ -26,7 +26,7 @@ function decodeUplink(bytes) {
 
 // type: 0x4 Registration
 function decodeRegistration(bytes) {
-  var data = {};
+  var data: any = {};
   data.type = "Registration";
   // adr
   data.adr = ((bytes[0] >> 3) & 0x1) == 0 ? "OFF" : "ON";
@@ -98,7 +98,7 @@ function decodeRegistration(bytes) {
 
 // type: 0x2 Heartbeat
 function decodeHeartbeat(bytes) {
-  var data = {};
+  var data: any = {};
   // type
   data.type = "Heartbeat";
   // battery
@@ -126,7 +126,7 @@ function decodeHeartbeat(bytes) {
 
 // type: 0x6 Beacon
 function decodeBeacon(bytes) {
-  var data = {};
+  var data: any = {};
   data.type = "Beacon";
   data.length = bytes[0] & 0x0f;
   for (let i = 0; i < data.length; i++) {
@@ -152,7 +152,7 @@ function decodeBeacon(bytes) {
 
 // type: 0xe AlarmBeaconList
 function decodeAlarmBeaconList(bytes) {
-  var data = {};
+  var data: any = {};
   data.type = "AlarmBeaconList";
   data.length = bytes[0] & 0x0f;
   for (let i = 0; i < data.length; i++) {
@@ -178,7 +178,7 @@ function decodeAlarmBeaconList(bytes) {
 
 // type: 0xf Acknowledgment
 function decodeAcknowledgment(bytes) {
-  var data = {};
+  var data: any = {};
   data.type = "Acknowledgment";
   data.result = (bytes[0] & 0x0f) == 0 ? "Success" : "Failure";
   data.msgId = (bytes[1] & 0xff).toString(16).toUpperCase();
@@ -186,11 +186,10 @@ function decodeAcknowledgment(bytes) {
   return data;
 }
 
-// payload
-var ignore_vars = [];
+var ignore_vars: any = [];
 
 function toTagoFormat(object_item, group, prefix = "") {
-  const result = [];
+  const result: any = [];
   for (const key in object_item) {
     if (ignore_vars.includes(key)) continue;
